chore: rename pox-2 file
