+++ src/event-stream/pox-event-parsing.ts
